chore: remove integration tests
